requirements.txt
numpy
scipy
librosa
torchmetrics
pytorch-lightning
pandas
matplotlib
scikit-learn
mpi4py
joblib
smplx
transforms3d
x-transformers
madmom
ftfy
axial_positional_embedding
einops
entmax
lightning-bolts
loguru
madgrad
mido
jsmin
optuna
